تشغيل حقيقي على PostgreSQL: 72/72 اختباراً ينجح
أول تشغيل فعلي بعد توفر Docker. ثلاثة أعطال حقيقية ظهرت وأُصلحت:

1. seed: upsert بمفتاح مركّب يرفض null — منصب CEO بلا قسم
   الإصلاح: findFirst ثم create/update

2. seed: نفس العطل في user_territories — areaId فارغ يعني
   "كل مناطق المحافظة"

3. تقرير التارغيت: DATE_PART يعيد double precision، و PostgreSQL
   لا يملك ROUND(double, int) — أُضيف ::int

تحقق فعلي على قاعدة بيانات حيّة:
- 31 جدولاً + pg_trgm + unaccent
- 86 حساباً · 85 رابطاً هرمياً · 89 إسناد منطقة · 13 منصباً
- 21 منتجاً · 138 اسماً بديلاً · 37 إذناً على 6 أدوار
- الدخول بالرقم وباللاحقة · تدوير التوكن يُبطل القديم
- البحث: "احمد" يجد "أحمد محمّد" — pg_trgm يعمل
- DR-00001/PH-00001 · اللقطة الـ15 حقلاً تُملأ فعلاً
- زيارة خارج القائمة تولّد عميلاً + طلب موافقة
- التقارير الـ11 تُنفَّذ · CTE الهرمية: 21 تابعاً بعمق 2
- Excel حيّ: 13 ورقة · 30 عموداً · جدول مسجَّل

أُضيف scripts/e2e-smoke.mjs — 72 حالة اختبار

+++ apps/api/src/modules/reports/reports.service.ts
@@ -249,10 +249,12 @@ export class ReportsService {
     const from = this.parseDate(range.from, 'تاريخ البداية');
     const to = this.parseDate(range.to, 'تاريخ النهاية');
 
-    // عدد الأشهر في الفترة — التارغيت شهري فيُضرب بها
+    // عدد الأشهر في الفترة — التارغيت شهري فيُضرب بها.
+    // ::int ضروري: DATE_PART يعيد double precision، و PostgreSQL
+    // لا يملك ROUND(double, int) — فقط ROUND(numeric, int).
     const months = Prisma.sql`
       GREATEST(1, (DATE_PART('year', ${to}::date) - DATE_PART('year', ${from}::date)) * 12
-                + (DATE_PART('month', ${to}::date) - DATE_PART('month', ${from}::date)) + 1)
+                + (DATE_PART('month', ${to}::date) - DATE_PART('month', ${from}::date)) + 1)::int
     `;
 
     return this.prisma.$queryRaw<any[]>`
